Rewrite SMTSolver rule and SMTInterface.
Functionality:
The solver can now handle all expressions which have some arithmetic or boolean operations.
It simply declares new variables for all subterms it cannot understand (such as pairs,
functions, function applications etc).

Design:
The work is split between SMTSolver rule, SMTInterface and SMTWriter. SMTSolver rule
finds equalitites in the context, collects the subtrees which should be declared as
the smt solver variables (actual variables and the subtrees whose structure the solver
cannot understand, as described above) and calls SMTInterface to declare the variables,
add all the assertions (arithmetic equalities as well as negation of the equality goal)
and perform the satisfiability check.
SMTInterface translates the assertions from StainlessFit trees to scala-smtlib terms
using variables, passed from the SMTSolver rule, for unknown subtrees, and calls
SMTWriter to write the declared variables and the assertions to a file.